changes to delete pages and replist for mobile view and qof changes

+++ workout_tracker/workouts/views.py
@@ -1,17 +1,19 @@
 from django.shortcuts import render, redirect, get_object_or_404, reverse
+from django.urls import reverse_lazy
 from .forms import *
 from .models import Workout_Split, Weight_and_reps
 from django.http import HttpResponseForbidden
-from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
-@login_required
 def delete_view(request):
+    if not request.user.is_authenticated:
+        return redirect(reverse_lazy('login'))
+    
     models = Workout_Split.objects.filter(user=request.user)
     if request.method == 'POST':
-        workout_id = request.POST.get('deleted_workout')
-        model = Workout_Split.objects.get(id=workout_id)
+        workout_id = request.POST.get('deleted_workout') # get workout user chose to delete
+        model = Workout_Split.objects.get(id=workout_id) # and delete it
         
         model.delete()
         return redirect('workoutsplitshowerview')
@@ -20,9 +22,11 @@ def delete_view(request):
     content['models'] = models
     return render(request,'delete_workout.html', content)
 
-@login_required
 def delete_day_view(request,pk):
-    my_model = Workout_Split.objects.get(id=pk)
+    if not request.user.is_authenticated:
+        return redirect(reverse_lazy('login'))
+    
+    my_model = Workout_Split.objects.get(id=pk) # check if its the users page and if not then block
     if my_model.user != request.user:
         return HttpResponseForbidden("You are not allowed to access this page.")
     
@@ -35,8 +39,10 @@ def delete_day_view(request,pk):
     workouts = Weight_and_reps.objects.filter(model=my_model).order_by('-date')
     return render(request,'delete_day.html', {'workouts': workouts})
 
-@login_required
 def calculator_view(request):
+    if not request.user.is_authenticated:
+        return redirect(reverse_lazy('login'))
+    
     form = Max_Rep_Calculator_Form(request.POST)
     onerep=-1
     
@@ -52,8 +58,10 @@ def calculator_view(request):
     return render(request, 'calculator.html', content)
 
 
-@login_required
 def workout_chooser(request):
+    if not request.user.is_authenticated:
+        return redirect(reverse_lazy('login'))
+    
     form = Workout_SplitForm(request.POST or None)
 
     if form.is_valid(): # adds user to the workout
@@ -68,22 +76,23 @@ def workout_chooser(request):
     return render(request, 'workout_chooser.html', context)
 
 
-@login_required
 def workoutsplitshowerview(request):
     workoutsplit = Workout_Split.objects.filter(user=request.user)
     return render(request, 'workoutsplitshower.html',{'workoutsplit':workoutsplit})
 
 
-@login_required
 def reps_view(request, pk):
+    if not request.user.is_authenticated:
+        return redirect(reverse_lazy('login'))
+    
     my_model = get_object_or_404(Workout_Split, pk=pk)
     if my_model.user != request.user:
         return HttpResponseForbidden("You are not allowed to access this page.")
     
-    if Weight_and_reps.objects.filter(model=my_model).exists():
+    if Weight_and_reps.objects.filter(model=my_model).exists(): # check if the user has workouts made
         weight_reps_model = Weight_and_reps.objects.filter(model=my_model).latest('date')
     else:
-        weight_reps_model = Weight_and_reps(model=my_model)
+        weight_reps_model = Weight_and_reps(model=my_model) # if not then create one
     
     form = Weight_and_reps_form(request.POST)
     if request.method == 'POST':
@@ -97,7 +106,7 @@ def reps_view(request, pk):
                 if onerepmax > weight_reps_model.max_weight:
                     weight_reps_model.max_weight = round(onerepmax,1)
                     
-                if weight_reps_model.reps != "":
+                if weight_reps_model.reps != "": # add to the text box of the model (cannot use array because not supported with sqlite)
                     weight_reps_model.reps += " "
                     weight_reps_model.weight += " "
                 weight_reps_model.reps += str(reps) + ","
